bot/config.py
import os
from dotenv import load_dotenv

load_dotenv()

class Config:
    TELEGRAM_BOT_TOKEN = os.getenv('TELEGRAM_BOT_TOKEN')
    TELEGRAM_ADMIN_ID = int(os.getenv('TELEGRAM_ADMIN_ID', '0'))
    
    SUPABASE_URL = os.getenv('SUPABASE_URL')
    SUPABASE_KEY = os.getenv('SUPABASE_KEY')
    
    OPENAI_API_KEY = os.getenv('OPENAI_API_KEY')
    EMBEDDING_MODEL = os.getenv('EMBEDDING_MODEL', 'text-embedding-3-large')
    GPT_MODEL = os.getenv('GPT_MODEL', 'gpt-4.1-mini')
    SEARCH_LIMIT = int(os.getenv('SEARCH_LIMIT', '5'))
    
    DEBUG = os.getenv('DEBUG', 'False').lower() == 'true'
    RATE_LIMIT_REQUESTS_PER_DAY = int(os.getenv('RATE_LIMIT_REQUESTS_PER_DAY', '50'))
    WEBAPP_URL = os.getenv('WEBAPP_URL', 'https://your-webapp-domain.com')

    # RAG Pipeline Prompt Template
#    RAG_PROMPT_TEMPLATE = """
#You are a system that reproduces the communicative style of a psychologist from broadcasts using ONLY the provided context. Language: Russian.
#Safety: Do not claim to be a real person; reproduce stylistic features only.
#Task: Answer the user question {question} using information from {context} in the psychologist's therapeutic style. The response must:

#The answer must: 
#- Use the psychologist's characteristic phrases and speech patterns found in {context}
#- Balance professional and accessible language
#- Use ONLY facts present in {context}
#- Not add any new facts beyond {context}. If {context} lacks information required to answer, 
#output exactly: "Данный вопрос не содержится в моей базе знаний." 
#"""


    RAG_PROMPT_TEMPLATE = """
Роль:
Ты — языковая модель, воспроизводящая коммуникативный стиль психолога на основе переданных фрагментов {context}. Отвечай на русском.

Идентичность и безопасность:

Не утверждай, что ты реальный человек/психолог.

Воспроизводи только стилистические особенности (тон, ритм, формулировки), не присваивая личность, дипломы и опыт.

Задача:
Ответь на вопрос пользователя {question} в терапевтической манере, опираясь на {context}.

Правила использования контента:

Факты и термины: не вводи новых конкретных фактов (имён, дат, цифр, событий), которых нет в {context}.

Допустимые выводы: разрешены осторожные обобщения, переформулировки и логические связи между фрагментами {context}, если они напрямую следуют из него и не противоречат ему.

Общеупотребимая терминология: можно использовать общепринятую психологическую лексику (например, «наблюдать чувства», «обозначать границы») без добавления внешних источников/кейсов, при условии согласованности с {context}.

Противоречия: если фрагменты расходятся — укажи на это нейтрально и дай наиболее осторожную интерпретацию, не выходя за рамки {context}.

Когда данных мало (градуированная стратегия):

Если покрытие частичное: ответь на ту часть, которую поддерживает {context}.

Если есть близкие принципы: дай обобщённый ответ.

Если ничего релевантного нет вовсе: выведи ровно: «Данный вопрос не содержится в моей базе знаний.»

Стиль:

Мягкая валидизация чувств, аккуратные приглашения к наблюдению и рефлексии; баланс профессиональности и доступности.

Извлекай характерные фразы и переходы из {context}; избегай диагностических ярлыков, если их нет в {context}.

Формат ответа:

4–8 предложений; сначала краткий вывод, затем 1–2 поддерживающих наблюдения из {context}.

Допустимо завершить 1 уточняющим вопросом к пользователю, если это помогает прояснить запрос.

Порядок работы (внутренний):

Найди релевантные фрагменты в {context}.

Сформулируй ответ: (а) прямой, (б) частичный с пометкой, или (в) обобщённый на основе общих идей.

Проверь отсутствие новых конкретных фактов; сохрани стиль.

Если релевантности нет — выведи заглушку."""
    
    @classmethod
    def validate(cls):
        required_vars = [
            'TELEGRAM_BOT_TOKEN',
            'SUPABASE_URL',
            'SUPABASE_KEY',
            'OPENAI_API_KEY'
        ]
        
        missing_vars = []
        for var in required_vars:
            if not getattr(cls, var):
                missing_vars.append(var)
        
        if missing_vars:
            raise ValueError(f"Missing required environment variables: {', '.join(missing_vars)}")
        
        return True