requirements.txt
aiogram>=3.0.0
python-dotenv
jinja2
python-multipart
langchain
langchain-openai
langchain-community
openai>=1.0.0
anthropic>=0.5.0
psycopg2-binary
faster-whisper
aiohttp
tlgbotfwk
langsmith 
pydantic
supabase>=2.0.0
postgrest>=0.10.0
requests>=2.31.0
elevenlabs>=0.2.26
argparse
pathlib
typing
